+++ library/core/src/main/java/com/google/android/exoplayer2/source/ClippingMediaSource.java
@@ -342,10 +342,8 @@ public final class ClippingMediaSource extends CompositeMediaSource<Void> {
     }
 
     @Override
-    public Window getWindow(
-        int windowIndex, Window window, boolean setTag, long defaultPositionProjectionUs) {
-      timeline.getWindow(
-          /* windowIndex= */ 0, window, setTag, /* defaultPositionProjectionUs= */ 0);
+    public Window getWindow(int windowIndex, Window window, long defaultPositionProjectionUs) {
+      timeline.getWindow(/* windowIndex= */ 0, window, /* defaultPositionProjectionUs= */ 0);
       window.positionInFirstPeriodUs += startUs;
       window.durationUs = durationUs;
       window.isDynamic = isDynamic;