+++ RELEASENOTES.md
@@ -40,6 +40,7 @@
   ([#5619](https://github.com/google/ExoPlayer/issues/5619)).
 * Fix issue where player errors are thrown too early at playlist transitions
   ([#5407](https://github.com/google/ExoPlayer/issues/5407)).
+* Deprecate `setTag` parameter of `Timeline.getWindow`. Tags will always be set.
 
 ### 2.10.4 ###